Move domain objects to correct package alongside where they are used.
