+++ tessera-jaxrs/sync-jaxrs/src/test/java/com/quorum/tessera/p2p/recovery/MockLegacyResendManager.java
@@ -1,8 +1,8 @@
 package com.quorum.tessera.p2p.recovery;
 
+import com.quorum.tessera.recovery.resend.ResendRequest;
+import com.quorum.tessera.recovery.resend.ResendResponse;
 import com.quorum.tessera.recovery.workflow.LegacyResendManager;
-import com.quorum.tessera.transaction.ResendRequest;
-import com.quorum.tessera.transaction.ResendResponse;
 
 public class MockLegacyResendManager implements LegacyResendManager {
