+++ tessera-recover/src/main/java/com/quorum/tessera/recovery/resend/ResendRequest.java
@@ -1,4 +1,4 @@
-package com.quorum.tessera.transaction;
+package com.quorum.tessera.recovery.resend;
 
 import com.quorum.tessera.data.MessageHash;
 import com.quorum.tessera.encryption.PublicKey;
