+++ tessera-recover/src/main/java/com/quorum/tessera/recovery/workflow/LegacyResendManagerImpl.java
@@ -9,8 +9,8 @@ import com.quorum.tessera.enclave.EncodedPayload;
 import com.quorum.tessera.enclave.PayloadEncoder;
 import com.quorum.tessera.enclave.PrivacyMode;
 import com.quorum.tessera.encryption.PublicKey;
-import com.quorum.tessera.transaction.ResendRequest;
-import com.quorum.tessera.transaction.ResendResponse;
+import com.quorum.tessera.recovery.resend.ResendRequest;
+import com.quorum.tessera.recovery.resend.ResendResponse;
 import com.quorum.tessera.transaction.exception.EnhancedPrivacyNotSupportedException;
 import com.quorum.tessera.transaction.exception.TransactionNotFoundException;
 import com.quorum.tessera.transaction.publish.PayloadPublisher;
